testserver: Use server factory to construct server
With commit 4971502f7 (server: Use injected Cache as cache invalidator,
2021-05-12), a segfault was introduced into Gitaly because the injected
cache wasn't actually set up as a field for the server factory. While
the segfault was easy to fix, it shows that we are lacking test coverage
of production code. This lacking coverage is caused by the fact that we
never construct servers for our tests via the server factory, but in
production we do.

Let's plug this blind spot by changing the testserver code to use the
server factory.