repository: Drop rename of attributes path
When syncing the repository's gitattributes file, we use the "safe"
package to safely write the new file and then atomically rename it into
place. But weirdly enough, `syncInfoAttributes()` is still doing a final
rename of the written gitattributes file to itself, which should be in
fact be a no-op. This has been like this ever since its original been
introduced in ab7a446c2 (Sync info attributes in ReplicateRepository,
2019-12-11). My assumption is that this is an artifact from code review,
where the MR might have been rewritten to make use of our "safe"
package, but the initial rename was still in place.

In any case, the rename shouldn't be necessary, so let's drop it.