repository: Split out logic to write repo files
With the upcoming introduction of synchronization for the gitconfig,
we're about to add a second callsite which wants to safely write a file
into a repository like the gitattributes synchronization currently does.
To prepare for this and avoid needless code duplication, this commit
refactors the code and splits out a function which writes the file into
the repo via the `safe` package.