repository: Replicate gitconfig
When replicating repositories, we do not currently replicate the
gitconfig. This is quite important to do though given that the gitconfig
may influence various different aspects of Git and may thus cause
mismatching results if it is different across replicas. Even if we
eventually remove the `AddConfig()` and `DeleteConfig()` RPCs, we'd
still want to replicate it given that different Git versions may end up
with different defaults and thus with different gitconfig files.

Implement replication of gitconfig files. As the `GetConfig()` RPC which
we require for this has only been introduced in this same commit series,
we need to play it safe and for now ignore all errors returned by it to
guard against the case where the remote Gitaly instance doesn't yet
support it. This workaround can be dropped in the next release cycle.

Changelog: added