repository: Enable transactional voting on the gitconfig
Whenever the gitconfig changes, we really ought to do transactional
voting to ensure that all replicas are running with the same gitconfig.
This feature is being introduced by this commit, where we always vote on
the complete gitconfig contents. Now that we have replication for this
file in place, it's safe enough to do and would eventually get repaired
if we notice divergent gitconfigs across replicas.

Changelog: added