ci: Fix code navigation job
Our code navigation CI job is currently broken, where we always get an
Internal Server Error when uploading artifact created by lsif-go. As it
turns out, this breakage is caused by an upstream change which has
caused a change in the output format.

Fix the error by pinning the lsif-go version to v1.3.1, which is the
most recent version which still works without issue.