coordinator: Add replication metrics for transactions
Without scraping through the logs, it's currently hard to determine how
many replication jobs we're queueing for transactional RPC and what the
reasons for those jobs are. To improve the situation, this commit adds a
new metric which tracks the reason why replication jobs are queued to
improve visibility.

Changelog: added