trace2: Expose child events to logs
While trace2 emits a lot of messages, this change disregards almost all
of them except Git shelling out to a child process. In that case we
capture only the most important details like the SID, arguments, and
exit code.

The trace2.CopyHandler has been renamed, and tests have been updated to
validate the messages in the logs.

Changelog: added