Remove isDone field from subtransaction type
isDone field in the subtransaction indicates whether the subtransaction
is already finished or not. It's set always when closing the doneCh, which
is used to signal to voters waiting for a subtransaction that it's finished.
Given the above, isDone really just duplicates the state of whether the doneCh
has been closed or not. This commit replaces isDone with a method that checks
doneCh's state to remove the duplication.