Makefile: Fix targets when running make(1) in an interactive rebase
During an interactive rebase, Git sets up various environment variables
which point to the repository in which the rebase is being executed.
Other Git commands will pick up these envvars and will as a result
operate in that same repository. This is dangerous in our case, given
that we're spawning Git commands in our Makefile which we certainly want
to run in a different directory.

Fix the issue by unexporting these environment variables.