transactions: Make logic to update voter states self-contained
Right now, updating the voter state is distributed across `vote()`,
`cancelSubtransaction()` and `updateVoterState()`. Given that this is
all non-trivial logic and that we have interdependencies between the
data that is to be updated, it should rather be located in a single
place.

As a first step towards this, move logic contained in `vote()` into
`updateVoterState()`. This includes registration of the vote itself as
well as a check whether the voters state matches what we'd expect.