transactions: Merge `cancelVote()` into `updateVoterState()`
Both `updateVoterState()` and `cancelVote()` update state of voters. But
while the former has additional sanity checks and also keeps track of
updating fellow voters in case any threshold was reached, the latter
doesn't. This is a potential bug waiting to happen: when we update the
logic to update voter state and change conditions for signalling, then
it may be that both code sites now need to do the signalling. And in
fact, a subsequent commit will cause us to fail a vote if outstanding
votes do not suffice anymore to bring any of the current votes across
the threshold, which will require signalling even for the cancellation
case.

To avoid this potential pitfal and reduce code duplication, this commit
thus merges both `updateVoterState()` and `cancelVote()` as a
preparatory step.