Simplify PerRepositoryElector's test setup
PerRepositoryElector's test currently take in function that can be
used to assert the primary. It turns out we don't need this as we're
always just asserting the primary is one of a given set of nodes or that
there is no primary. Replace the function field with just a list of
acceptable primaries. This allows us later to use the information of which
primaries are expected which would currently be unavailable as the information
is contained in the closure.