transactions: Decide signalling based on voter state
Whether we need to signal or not is currently, decide other things,
decided based on whether all voters have cast a vote or not. We have
recently started to support retracting votes though. While we leave the
nodes vote in place right now, it's thinkable that we'll eventually
unset it. And as soon as we do that, we may easily mislabel that node as
undecided.

Improve the code by deciding based on the voter's result, not on its
vote.