transactions: Fail early if the threshold cannot be reached anymore
When we have more than three nodes, then it is possible to determine
that a transaction cannot succeed anymore even when not all votes were
cast yet. If you for example have four nodes with a vote count of one
each, and a threshold of three, then it is always possible to decide the
outcome after you've recorded three votes. Waiting for the fourth node
to arrive is pointless in that case, and instead we can already move
ahead with the other nodes.

Implement this idea for our subtransactions.

Changelog: added