coordinator: Change order to compute updated/outdated secondaries
Shuffle around the order in which we compute updated and outdated
secondaries such that we have the `primaryDirtied` state available for
all of them. No functional change is expected from this commit, but it's
only preparing for the next commit.

(cherry picked from commit 30e482604548b0d923f13b11dbc5aa63fdc289f9)