Query the health consensus directly in PerRepositoryElector
PerRepositoryElector currently gets the health consensus from the
HealthManager. This is an unnecessary loop as the primary elector
could query for the consensus directly from the database. This was
originally implemented like this to ease testing and avoid duplicating
the query logic. With the query logic extracted into a view, let's
query the view directly.