Extract repository_generations and valid_primaries views
This commit extracts the logic for getting the valid primaries into a
view for easier reuse. This logic will be reused when implementing lazy
failovers in `praefect dataloss` and the read-only repository metric.
Both of them currently check whether the primary stored in the database
is outdated to determine whether the repository is in read-only mode or
not. This won't be sufficient with lazy failovers anymore. The repository
would only get a new primary if it is accessed while the current primary
is unhealthy and there are valid primary candidates available. As such,
both of these tools will need to check whether there are valid primaries
to failover before denoting the repository as read-only.

The repository_generations view is also extracted and used in the valid_primaries
view. Many queries currently check the storage_repositories records to figure
out what is the latest generation for a given repository. The view deduplicates
this logic.