Prevent usage of other election strategies than per_repository
Praefect's legacy election strategies have been deprecated in 13.12
and are scheduled for removal in 14.0. This commit switches the
`per_repository` to be the default election strategy and ignores the
`failover.election_strategy` configuration key.

As there are still some tests in Gitaly's repo and in GitLab's repo
which are not configured to use the database, there's an additional
configuration key added which allows for still configuring the other
election strategies.

Changelog: removed