Write commit graph using bloom filter
Bloom filters speeds up execution of the 'git log'
and other commands that take a pathspec argument.
Let's enable Bloom filters in the commit graph by
passing the --changed-paths flag to git commit-graph.
If Bloom filters were not used for the existing commit graph
files we should remove them and re-create the commit graph
files as `--changed-paths` flag in this case would be skipped.
The --split=replace strategy retroactively computes Bloom
filters over the earlier layers and replaces the old commit graph
file or chain with a file set.
If commit graph creation is in progress the GarbageCollect call
will fail as it is not allowed to run multiple graph file
creation operations for the same repository. The operation uses
lock file to handle concurrent commit graph creation operations.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3545