Replace GetPartiallyReplicatedRepositories with GetPartiallyAvaialableRepositories
`praefect dataloss` is using GetPartiallyReplicatedRepositories to get
repositories which have assigned replicas that are outdated. Inferring from the
returned generations it was also reporting whether the repository was in read-only
mode or not. This is not sufficient anymore to determine whether a repository is
unavailable or not due to recent changes:

1. Since 7af9c950, Praefect has no longer served reads from outdated replicas.

2. Praefect no longer elects outdated replicas as primaries. Electing an outdated
   primary does not improve the availability of a repository as it still couldn't
   accept writes nor reads.

3. With introduction of lazy failovers, there is effectively no read-only mode
   anymore as Praefect would simply failover to the up to date node immediately
   if one exists.

With those in mind, the behavior of `praefect dataloss` is not accurate anymore.
By default, its attempts to print out repositories which have reduced availability.
To reflect the current failover logic, we should instead print out repositories
which do not have any up to date, healthy nodes available. This commit replaces
the GetPartiallyReplicatedRepositories with GetPartiallyAvailableRepositories.
A repository is considered available by the current logic if there exists a replica
that could serve as the primary. A replica can serve as the primary if it is fully
up to date and healthy. If such a replica exists, the repository is not in read-only
mode as we'd simply use the replica as the primary. If no such replicas exist, the
repository is unavailable.

The dataloss sub-command also has the `-partially-replicated` flag that prints out
repositories which have some assigned replicas that are not fully up to date. That
flag is going to be replaced by the `partially-available` flag, which returns
repositories which have assigned replicas that are not able to serve requests at
the moment. This effectively does the same as the flag did previously but it also
considers whether the replicas are healthy. This behavior fits better with variable
replication factor: it could be that we have one up to date copy of the replica on
an unhealthy node. The previous check would only see that there are no outdated
replicas and not return the repository. The repository would be unavailable though,
as the only replicas is on a node that is unhealthy. To better facilitate debugging
these scenarios, the flag is changed to cover replicas on unavailable nodes as well.

This commit covers only the datastore changes. The user facing changes in dataloss
are to be done in a follow up commit.