nodes: Mention gitaly in 'ErrPrimaryNotHealthy'
Gitaly errors messages will be shown to end-users in the web UI.
Without knowing that the error message for 'ErrPrimaryNotHealthy' is
from Praefect, it's difficult for users to understand what 'primary' is
referring to.

Let's mention that Gitaly is the source of the failure to make this less
confusing.

Changelog: changed