Optimize valid_primaries view
The valid_primaries view contains storages which are eligible to act
as primaries for repositories. A replica is considered to be a valid
primary if:

1. It is healthy by the consensus of Praefect nodes.
2. It is fully up to date.
3. It is not pending deletion.
4. It is assigned, or there are no assigned replicas which are considered
   to be valid primaries.

The query is currently implemented in a suboptimal way which causes the 4th
check to kill the performance completely. The root of the problem is the
candidates CTE which is used twice in the view's query. Postgres's query optimizer
has a limitation where it's not able to push the parent query's restrictions
into the CTE query if it is referenced more than once. Combined with the 4th check,
this causes the CTE to be queried once for all of the candidates. The query times blow
up as a result.

This commit optimizes the view in the query by removing the CTE entirely and using
a subquery instead. This allows the Postgres to better optimize the query. Now it doesn't
have to compute the whole CTE query for all of the records when fetching records of a single
repository. Likewise, it no longer loops over the CTE once for each candidate. The
performance improves drastically. Testing in the 10k reference architecture environment,
querying for 10000 valid primaries took originally 55 seconds. With the optimizations in
place, the same query takes only 257 milliseconds.