localrepo: Refactor `Push()` tests to allow more control over test repo
We're about to add the ability to pass Git configuration entries to
`localrepo.Push()`. As the intended use case for this is to allow for
in-memory remotes when pushing, one of the new testcases will require
the ability to more tighlty control the push repository setup.

Refactor the test code to prepare for this by always having testcases
setup their own repositories.