proto: Use Go identifiers for UpdateRemoteMirror documentation
When generating Go code from protobuf messages, we always have a
mismatch between declared variables and generated identifiers. This
leads to a mismatch in generated documentation, where the variable
documentation does not start with the Go variable name as it is best
practice in Go. Because of this, we have started to use Go identifiers
in protobuf documentation to fix the discrepancy.

Update documentation of variables in messages related to
UpdateRemoteMirror to do so consistenly.