remote: Allow for in-memory remotes in UpdateRemoteMirror
As we're about to phase out usage of on-disk remotes, all RPCs which
interact with them need to be converted to use in-memory remotes.
`UpdateRemoteMirror()` doesn't yet have the ability to use them.

Implement support for in-memory remotes by configuring the remote via a
set of Git configuration environment variables. Until downstream callers
have been converted from on-disk to in-memory remotes, both ways will
stay supported. Afterwards, we can drop the `RefName` parameter, which
really is the remote name.

Changelog: added