repository: Relax URL check when fetching remotes
The `FetchRemote()` RPC is used to fetch from a remote into the target
repository. Traditionally, this has been used with an on-disk remote,
where a previous call to `AddRemote()` was requiret to set up a remote
with an associated Git URL. This has changed with the introduction of
the `Remote` parameter, which specifies an in-memory remote with URL and
optionally a set of credentials. With this, the caller does not need to
call `AddRemote()` anymore and we have to rely less on on-disk state.

There is a major discrepancy between those two calls though which makes
it infeasible to use in-memory remotes for all usecases which currently
exist downstream in Rails: while `AddRemote()` would accept all URLs,
the in-memory URL is getting validated via `url.ParseRequestURI()` and
requires a valid host. This is overly strict: first it fails to parse
SSH URIs (e.g. "git@example.com:path/to/repo.git"), and second file
paths do not have a valid host. Furthermore, this validation has only
been implemented in the Go port. The old Ruby port removed in d1a384ffd
(Removal of ruby implementation of the FetchRemote, 2020-12-28) didn't
verify the URL at all.

This hasn't yet bitten us because downstream only uses in-memory remotes
in some places. Most notably, Rails is still using on-disk remotes for
`fetch_upstream()`. While users may only specify a limited set of Git
URLs with a proper schema, tests assert that it's also possible to fetch
from a file path directly.

Let's relax the URL check and simply assert that it's not empty. Rails
should be responsible for enforcing that the user specifies a valid URL,
and we shouldn't restrict valid URLs.

Changelog: changed