Make gitaly supervisor configuration local
Supervisor for the gitaly-ruby sidecar uses a package global
variable to store configuration. That causes issues in testing
as it produces a race conditions some times. The global variables
is an evil and we try to get rid of them. That change removes
the global variable and adds a constructor for the Config type.
Each test now uses a locally defined config variable and there is
no races on the global variable anymore.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3651