blackbox: Make sleep duration private
We have two representations of the sleep duration in the blackbox
Config: once a normal integer which gets parsed from the config file,
and once the converted `time.Duration`. Ideally, we'd have used a
`time.Duration` from the inception such that it's possible to specify
human-readable durations in the config file and not require any
conversion, but that train has departed by now given that the
gitaly-blackbox is already in use.

While we cannot merge those two variables now, we can at least make the
`SleepDuration` private to the package as it's not used anywhere else.