blackbox: Pass around configuration as value
We typically pass configuration structures around as a value instead of
as a pointer to make the configuration immutable to the receiving side.
One exemption to thi sis the blackbox code, where we pass it around as a
pointer.

Bring the code more in line with our own best practices and pass it as
simple value.