blackbox: Implement logic on a Blackbox struct receiver
Right now, the implenetation of Blackbox logic always gets a
configuration passed in as it is not associated to any structure. This
shortcoming keeps us from having all logic like for example also the
Prometheus counters as part of the struct, which means that they
currently live as globals.

Let's prepare to get rid of these globals by creating a new `Blackbox`
structure and implement blackbox logic as receiver functions. For now,
the struct only hosts the configuration.