stats: Let callers append newlines for `printInteractive()`
When `printInteractive()` is called, the assumption is that all lines
should be terminated by a newline. While that's true right now, this
assumption keeps us from using this function in other places which thus
need to manually check whether the `interactive` flag is set.

Do not append the newline in `printInteractive()` and adjust callers
accordingly to prepare for wider use of that function.