stats: Split out fetch pack parsing
The code to parse the git-fetch-pack(1) body is tightly coupled with the
code which drives the actual request. This makes it hard to reason about
both concerns and also makes it hard to just parse a fetch pack body
without doing the actual request.

Refactor the code and split out a new `FetchPack` structure which
contains all logic to parse from an `io.Reader` without having to care
about the request itself. This structure is the equivalent to the
already existing `ReferenceDiscovery` structure.